requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
google-generativeai
